Addressed @subhasis' comments
