+++ src/cmds/Makefile.am
@@ -88,20 +88,19 @@ dist_bin_SCRIPTS = \
 
 common_cflags = \
 	-I$(top_srcdir)/src/include \
-	@KRB5_CFLAGS@ \
-	@libz_inc@
+	@KRB5_CFLAGS@
 
 common_libs = \
 	$(top_builddir)/src/lib/Libpbs/.libs/libpbs.a \
-	$(top_builddir)/src/lib/Libtpp/libtpp.a \
-	$(top_builddir)/src/lib/Liblog/liblog.a \
-	$(top_builddir)/src/lib/Libutil/libutil.a \
 	$(top_builddir)/src/lib/Libnet/libnet.a \
 	$(top_builddir)/src/lib/Libsec/libsec.a \
+	$(top_builddir)/src/lib/Libutil/libutil.a \
 	-lpthread \
 	@socket_lib@ \
-	@libz_lib@ \
 	@KRB5_LIBS@
+
+common_sources = \
+	$(top_srcdir)/src/lib/Libcmds/cmds_common.c
 
 pbsdsh_CPPFLAGS = ${common_cflags}
 pbsdsh_LDADD = ${common_libs}
@@ -109,11 +108,11 @@ pbsdsh_SOURCES = pbsdsh.c
 
 pbsnodes_CPPFLAGS = ${common_cflags}
 pbsnodes_LDADD = ${common_libs}
-pbsnodes_SOURCES = pbsnodes.c
+pbsnodes_SOURCES = pbsnodes.c ${common_sources}
 
 pbs_attach_CPPFLAGS = ${common_cflags}
 pbs_attach_LDADD = ${common_libs}
-pbs_attach_SOURCES = pbs_attach.c
+pbs_attach_SOURCES = pbs_attach.c ${common_sources}
 
 pbs_demux_CPPFLAGS = ${common_cflags}
 pbs_demux_LDADD = ${common_libs}
@@ -128,12 +127,11 @@ pbs_ds_password_bin_LDADD = \
 	@database_lib@ \
 	-lssl \
 	-lcrypto
-
-pbs_ds_password_bin_SOURCES = pbs_ds_password.c
+pbs_ds_password_bin_SOURCES = pbs_ds_password.c ${common_sources}
 
 pbs_password_CPPFLAGS = ${common_cflags}
 pbs_password_LDADD = ${common_libs}
-pbs_password_SOURCES = pbs_password.c
+pbs_password_SOURCES = pbs_password.c ${common_sources}
 
 pbs_tmrsh_CPPFLAGS = ${common_cflags}
 pbs_tmrsh_LDADD = ${common_libs}
@@ -141,47 +139,47 @@ pbs_tmrsh_SOURCES = pbs_tmrsh.c
 
 pbs_ralter_CPPFLAGS = ${common_cflags}
 pbs_ralter_LDADD = ${common_libs}
-pbs_ralter_SOURCES = pbs_ralter.c cmds_common.c
+pbs_ralter_SOURCES = pbs_ralter.c ${common_sources}
 
 pbs_rdel_CPPFLAGS = ${common_cflags}
 pbs_rdel_LDADD = ${common_libs}
-pbs_rdel_SOURCES = pbs_rdel.c
+pbs_rdel_SOURCES = pbs_rdel.c ${common_sources}
 
 pbs_rstat_CPPFLAGS = ${common_cflags}
 pbs_rstat_LDADD = ${common_libs}
-pbs_rstat_SOURCES = pbs_rstat.c
+pbs_rstat_SOURCES = pbs_rstat.c ${common_sources}
 
 pbs_rsub_CPPFLAGS = ${common_cflags}
 pbs_rsub_LDADD = ${common_libs}
-pbs_rsub_SOURCES = pbs_rsub.c cmds_common.c
+pbs_rsub_SOURCES = pbs_rsub.c ${common_sources}
 
 pbs_migrate_users_CPPFLAGS = ${common_cflags}
 pbs_migrate_users_LDADD = ${common_libs}
-pbs_migrate_users_SOURCES = pbs_migrate_users.c
+pbs_migrate_users_SOURCES = pbs_migrate_users.c ${common_sources}
 
 pbs_release_nodes_CPPFLAGS = ${common_cflags}
 pbs_release_nodes_LDADD = ${common_libs}
-pbs_release_nodes_SOURCES = pbs_release_nodes.c
+pbs_release_nodes_SOURCES = pbs_release_nodes.c ${common_sources}
 
 qalter_CPPFLAGS = ${common_cflags}
 qalter_LDADD = ${common_libs}
-qalter_SOURCES = qalter.c cmds_common.c
+qalter_SOURCES = qalter.c ${common_sources}
 
 qdel_CPPFLAGS = ${common_cflags}
 qdel_LDADD = ${common_libs}
-qdel_SOURCES = qdel.c
+qdel_SOURCES = qdel.c ${common_sources}
 
 qdisable_CPPFLAGS = ${common_cflags}
 qdisable_LDADD = ${common_libs}
-qdisable_SOURCES = qdisable.c
+qdisable_SOURCES = qdisable.c ${common_sources}
 
 qenable_CPPFLAGS = ${common_cflags}
 qenable_LDADD = ${common_libs}
-qenable_SOURCES = qenable.c
+qenable_SOURCES = qenable.c ${common_sources}
 
 qhold_CPPFLAGS = ${common_cflags}
 qhold_LDADD = ${common_libs}
-qhold_SOURCES = qhold.c
+qhold_SOURCES = qhold.c ${common_sources}
 
 qmgr_CPPFLAGS = \
 	${common_cflags} \
@@ -189,56 +187,56 @@ qmgr_CPPFLAGS = \
 qmgr_LDADD = \
 	${common_libs} \
 	@editline_lib@
-qmgr_SOURCES = qmgr.c cmds_common.c
+qmgr_SOURCES = qmgr.c ${common_sources}
 
 qmove_CPPFLAGS = ${common_cflags}
 qmove_LDADD = ${common_libs}
-qmove_SOURCES = qmove.c
+qmove_SOURCES = qmove.c ${common_sources}
 
 qorder_CPPFLAGS = ${common_cflags}
 qorder_LDADD = ${common_libs}
-qorder_SOURCES = qorder.c
+qorder_SOURCES = qorder.c ${common_sources}
 
 qmsg_CPPFLAGS = ${common_cflags}
 qmsg_LDADD = ${common_libs}
-qmsg_SOURCES = qmsg.c
+qmsg_SOURCES = qmsg.c ${common_sources}
 
 qrerun_CPPFLAGS = ${common_cflags}
 qrerun_LDADD = ${common_libs}
-qrerun_SOURCES = qrerun.c
+qrerun_SOURCES = qrerun.c ${common_sources}
 
 qrls_CPPFLAGS = ${common_cflags}
 qrls_LDADD = ${common_libs}
-qrls_SOURCES = qrls.c
+qrls_SOURCES = qrls.c ${common_sources}
 
 qrun_CPPFLAGS = ${common_cflags}
 qrun_LDADD = ${common_libs}
-qrun_SOURCES = qrun.c
+qrun_SOURCES = qrun.c ${common_sources}
 
 qselect_CPPFLAGS = ${common_cflags}
 qselect_LDADD = ${common_libs}
-qselect_SOURCES = qselect.c
+qselect_SOURCES = qselect.c ${common_sources}
 
 qsig_CPPFLAGS = ${common_cflags}
 qsig_LDADD = ${common_libs}
-qsig_SOURCES = qsig.c
+qsig_SOURCES = qsig.c ${common_sources}
 
 qstat_CPPFLAGS = ${common_cflags}
 qstat_LDADD = ${common_libs}
-qstat_SOURCES = qstat.c
+qstat_SOURCES = qstat.c ${common_sources}
 
 qstart_CPPFLAGS = ${common_cflags}
 qstart_LDADD = ${common_libs}
-qstart_SOURCES = qstart.c
+qstart_SOURCES = qstart.c ${common_sources}
 
 qstop_CPPFLAGS = ${common_cflags}
 qstop_LDADD = ${common_libs}
-qstop_SOURCES = qstop.c
+qstop_SOURCES = qstop.c ${common_sources}
 
 qsub_CPPFLAGS = ${common_cflags}
 qsub_LDADD = ${common_libs}
-qsub_SOURCES = qsub.c cmds_common.c
+qsub_SOURCES = qsub.c ${common_sources}
 
 qterm_CPPFLAGS = ${common_cflags}
 qterm_LDADD = ${common_libs}
-qterm_SOURCES = qterm.c
+qterm_SOURCES = qterm.c ${common_sources}
