+++ src/iff/Makefile.am
@@ -40,14 +40,11 @@ sbin_PROGRAMS = pbs_iff
 
 pbs_iff_CPPFLAGS = \
 	-I$(top_srcdir)/src/include \
-	@KRB5_CFLAGS@ \
-	@libz_inc@
+	@KRB5_CFLAGS@
 
 pbs_iff_LDADD = \
 	$(top_builddir)/src/lib/Libpbs/.libs/libpbs.a \
-	$(top_builddir)/src/lib/Libtpp/libtpp.a \
 	$(top_builddir)/src/lib/Libutil/libutil.a \
-	$(top_builddir)/src/lib/Liblog/liblog.a \
 	$(top_builddir)/src/lib/Libnet/libnet.a \
 	$(top_builddir)/src/lib/Libsec/libsec.a \
 	-lpthread \
@@ -55,4 +52,4 @@ pbs_iff_LDADD = \
 	@libz_lib@ \
 	@KRB5_LIBS@
 
-pbs_iff_SOURCES = iff2.c
+pbs_iff_SOURCES = iff2.c $(top_srcdir)/src/lib/Libcmds/cmds_common.c
