+++ src/lib/Libtpp/Makefile.am
@@ -43,6 +43,7 @@ libtpp_a_CPPFLAGS = \
 	@KRB5_CFLAGS@
 
 libtpp_a_SOURCES = \
+	tpp_internal.h \
 	tpp_client.c \
 	tpp_em.c \
 	tpp_platform.c \
