+++ src/tools/Makefile.am
@@ -59,18 +59,14 @@ EXTRA_PROGRAMS = \
 
 common_cflags = \
 	-I$(top_srcdir)/src/include \
-	@KRB5_CFLAGS@ \
-	@libz_inc@
+	@KRB5_CFLAGS@
 
 common_libs = \
 	$(top_builddir)/src/lib/Libpbs/.libs/libpbs.a \
-	$(top_builddir)/src/lib/Libtpp/libtpp.a \
-	$(top_builddir)/src/lib/Liblog/liblog.a \
 	$(top_builddir)/src/lib/Libutil/libutil.a \
 	$(top_builddir)/src/lib/Libnet/libnet.a \
 	$(top_builddir)/src/lib/Libsec/libsec.a \
 	-lpthread \
-	@libz_lib@ \
 	@socket_lib@ \
 	@KRB5_LIBS@
 
@@ -90,7 +86,7 @@ pbs_ds_monitor_LDADD = \
 	-lssl \
 	-lcrypto
 
-pbs_ds_monitor_SOURCES = pbs_ds_monitor.c
+pbs_ds_monitor_SOURCES = pbs_ds_monitor.c $(top_srcdir)/src/lib/Libcmds/cmds_common.c
 
 pbs_idled_CPPFLAGS = ${X_CFLAGS} ${common_cflags}
 pbs_idled_LDADD = \
@@ -98,7 +94,7 @@ pbs_idled_LDADD = \
 	${X_PRE_LIBS} \
 	${X_LIBS} \
 	-lX11
-pbs_idled_SOURCES = pbs_idled.c
+pbs_idled_SOURCES = pbs_idled.c $(top_srcdir)/src/lib/Libcmds/cmds_common.c
 
 pbs_hostn_CPPFLAGS = ${common_cflags}
 pbs_hostn_LDADD = ${common_libs}
@@ -120,6 +116,7 @@ pbs_python_CPPFLAGS =  \
 
 pbs_python_LDADD = \
 	$(top_builddir)/src/lib/Libpython/libpbspython.a \
+	$(top_builddir)/src/lib/Liblog/liblog.a \
 	$(top_builddir)/src/lib/Libattr/libattr.a \
 	${common_libs} \
 	@PYTHON_LDFLAGS@ \
@@ -137,15 +134,25 @@ pbs_python_SOURCES = \
 	$(top_srcdir)/src/server/resc_attr.c \
 	$(top_srcdir)/src/server/setup_resc.c \
 	$(top_srcdir)/src/server/vnparse.c \
+	$(top_srcdir)/src/lib/Libcmds/cmds_common.c \
 	pbs_python.c
 
 pbs_tclsh_CPPFLAGS = \
 	${common_cflags} \
-	@tcl_inc@ \
-	@libz_inc@
+	@libz_inc@ \
+	@tcl_inc@
 
 pbs_tclsh_LDADD = \
-	${common_libs} \
+	$(top_builddir)/src/lib/Libpbs/.libs/libpbs.a \
+	$(top_builddir)/src/lib/Libtpp/libtpp.a \
+	$(top_builddir)/src/lib/Liblog/liblog.a \
+	$(top_builddir)/src/lib/Libutil/libutil.a \
+	$(top_builddir)/src/lib/Libnet/libnet.a \
+	$(top_builddir)/src/lib/Libsec/libsec.a \
+	-lpthread \
+	@socket_lib@ \
+	@KRB5_LIBS@ \
+	@libz_lib@ \
 	@tcl_lib@
 
 pbs_tclsh_SOURCES = \
@@ -159,10 +166,20 @@ pbs_upgrade_job_SOURCES = pbs_upgrade_job.c
 
 pbs_wish_CPPFLAGS = \
 	${common_cflags} \
+	@libz_inc@ \
 	@tk_inc@
 
 pbs_wish_LDADD = \
-	${common_libs} \
+	$(top_builddir)/src/lib/Libpbs/.libs/libpbs.a \
+	$(top_builddir)/src/lib/Libtpp/libtpp.a \
+	$(top_builddir)/src/lib/Liblog/liblog.a \
+	$(top_builddir)/src/lib/Libutil/libutil.a \
+	$(top_builddir)/src/lib/Libnet/libnet.a \
+	$(top_builddir)/src/lib/Libsec/libsec.a \
+	-lpthread \
+	@socket_lib@ \
+	@KRB5_LIBS@ \
+	@libz_lib@ \
 	@tk_lib@
 
 pbs_wish_SOURCES = \
@@ -172,7 +189,7 @@ pbs_wish_SOURCES = \
 
 printjob_bin_CPPFLAGS = ${common_cflags}
 printjob_bin_LDADD = ${common_libs}
-printjob_bin_SOURCES = printjob.c
+printjob_bin_SOURCES = printjob.c $(top_srcdir)/src/lib/Libcmds/cmds_common.c
 
 printjob_svr_bin_CPPFLAGS = \
 	${common_cflags} \
@@ -188,7 +205,7 @@ printjob_svr_bin_LDADD = \
 	-lssl \
 	-lcrypto
 
-printjob_svr_bin_SOURCES = printjob.c
+printjob_svr_bin_SOURCES = printjob.c $(top_srcdir)/src/lib/Libcmds/cmds_common.c
 
 rstester_CPPFLAGS = ${common_cflags}
 rstester_LDADD = ${common_libs}
@@ -197,5 +214,6 @@ rstester_SOURCES = rstester.c
 tracejob_CPPFLAGS = ${common_cflags}
 tracejob_LDADD = ${common_libs}
 tracejob_SOURCES = \
+	$(top_srcdir)/src/lib/Libcmds/cmds_common.c \
 	tracejob.c \
 	tracejob.h
