+++ win_configure/projects/Libtpp.vcxproj
@@ -116,6 +116,9 @@
     <ClCompile Include="..\..\src\lib\Libtpp\tpp_util.c" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\..\src\lib\Libtpp\tpp_internal.h" />
+  </ItemGroup>
+  <ItemGroup>
     <ProjectReference Include="include.vcxproj">
       <Project>{0b889879-456f-4f8e-a15e-3c0a7bbf3b25}</Project>
       <ReferenceOutputAssembly>false</ReferenceOutputAssembly>
